feat: rename swebench example
